implemented paystack payment in payment activity
