+++ app/src/main/res/drawable/button_background.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/orange" />
+
+    <corners android:radius="5dp"/>
+
+</shape>
