+++ app/src/main/res/drawable/edit_text_background.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/iconicWhite" />
+    <stroke android:color="@color/orange"/>
+    <stroke android:width="2dp"/>
+    <corners android:radius="5dp" />
+</shape>
